LayoutTests/media/remove-from-document.html

<video controls></video>
<p>Test that removing a media element from the tree pauses playback but does not unload the media.</p>
<script src=video-test.js></script>
<script>
waitForEvent('load', function () {
    test("video.networkState == HTMLMediaElement.LOADED");
    test("!video.paused");
    run("document.body.removeChild(video)");
    test("video.networkState == HTMLMediaElement.LOADED");
    test("video.paused");
    document.body.offsetTop;
    endTest();
} );
video.src = "content/test.mp4";
video.play();
</script>
