LayoutTests/media/remove-from-document-expected.txt

Test that removing a media element from the tree pauses playback but does not unload the media.

EVENT(load)
TEST(video.networkState == HTMLMediaElement.LOADED) OK
TEST(!video.paused) OK
RUN(document.body.removeChild(video))
TEST(video.networkState == HTMLMediaElement.LOADED) OK
TEST(video.paused) OK
END OF TEST
